requirements.txt
attr==0.3.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.8
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
faiss_cpu==1.8.0
fitz==0.0.1.dev2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.3
Jinja2==3.1.4
jnius==1.1.0
keyring==25.2.1
llama_index==0.10.50
lockfile==0.12.2
mock==5.1.0
nltk==3.8.1
numpy==2.0.0
Pillow==9.5.0
Pillow==10.3.0
protobuf==5.27.1
pyOpenSSL==24.1.0
railroad==0.5.0
sentence_transformers==3.0.1
Sphinx==7.3.7
tensorflow==2.16.1
thread==2.0.3
tornado==6.4.1
trove_classifiers==2024.5.22
truststore==0.9.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
